# swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: FlukeBase Connect API
  version: v1
  description: API for flukebase_connect MCP server integration. Provides authentication,
    project management, environment variables, memories, and AI productivity metrics.
  contact:
    name: FlukeBase Support
    email: support@flukebase.com
paths:
...
  "/api/v1/flukebase_connect/projects":
    get:
      summary: List accessible projects
      tags:
      - Projects
      description: Returns all projects the authenticated user has access to. Includes
        projects the user owns and projects they collaborate on.
      operationId: listProjects
      security:
      - bearer_auth: []
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      "$ref": "#/components/schemas/project"
                required:
                - projects
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/error"
    post:
      summary: Create a new project
      tags:
      - Projects
      description: Creates a new project for the authenticated user.
      operationId: createProject
      security:
      - bearer_auth: []
      parameters: []
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/project"
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: My New Project
                description:
                  type: string
                  example: A description of the project
                repository_url:
                  type: string
                  example: https://github.com/user/repo
                stage:
                  type: string
                  enum:
                  - idea
                  - prototype
                  - mvp
                  - growth
                  - mature
                  example: prototype
              required:
              - name
  "/api/v1/flukebase_connect/projects/{id}":
    parameters:
    - name: id
      in: path
      description: Project ID
      required: true
      schema:
        type: integer
    get:
      summary: Get project details
      tags:
      - Projects
      description: Returns detailed information about a specific project.
      operationId: getProject
      security:
      - bearer_auth: []
      responses:
        '200':
          description: Project retrieved successfully
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/project"
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/error"
...
servers:
- url: https://flukebase.com
  description: Production server
- url: http://localhost:3006
  description: Development server
components:
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: API Token
      description: FlukeBase API token (starts with fbk_)
  schemas:
    error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
      - error
    project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        repository_url:
          type: string
          nullable: true
        stage:
          type: string
          enum:
          - idea
          - prototype
          - mvp
          - growth
          - mature
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - id
      - name
      - stage
...
security:
- bearer_auth: []